ft_transcendence/match_making/requirements.txt
Django==3.1
djangorestframework==3.14
djangorestframework-simplejwt
requests==2.31
channels==3.0
channels_redis==4.0
django-channels-presence==1.0
redis==5.0
django-redis==5.0
psycopg2==2.9.9

gunicorn 
uvicorn[standard]
httpx